CFIN-383: Refactor - move course search to new module
