+++ src/tests/pages/index.test.js
@@ -1,8 +1,11 @@
 import { render, screen } from "@testing-library/react";
 import App, { getServerSideProps } from "../../pages";
+import { searchForCourses } from "../../utils/searchForCourses";
+
+jest.mock("../../utils/searchForCourses");
 
 beforeEach(() => {
-    fetch.resetMocks();
+    jest.clearAllMocks();
 });
 
 const emptyContext = { query: {} };
@@ -41,75 +44,44 @@ describe("App", () => {
 });
 
 describe("getServerSideProps", () => {
-    it("calls the Courses API and returns response in expected format", async () => {
-        fetch.mockResponse(
-            JSON.stringify({
-                numberOfMatches: 1,
-                results: [
-                    {
-                        title: "English",
-                    },
-                    {
-                        title: "Maths",
-                    },
-                ],
-            })
-        );
+    it("calls course search and returns response in expected format", async () => {
+        searchForCourses.mockResolvedValue({
+            isSuccessfulSearch: true,
+            searchResponseData: { numberOfMatches: 1, results: [{ title: "English" }, { title: "Maths" }] },
+        });
 
         const response = await getServerSideProps(contextWithSearchTerm);
 
-        expect(fetch).toHaveBeenCalledTimes(1);
+        expect(searchForCourses).toHaveBeenCalledTimes(1);
 
         expect(response.props.isSuccessfulSearch).toEqual(true);
         expect(response.props.searchResults).toEqual([{ title: "English" }, { title: "Maths" }]);
         expect(response.props.searchTerm).toEqual("english");
     });
 
-    it("constructs the Courses API url with the expected environment variables", async () => {
-        await getServerSideProps(contextWithSearchTerm);
-
-        expect(fetch).toHaveBeenCalledTimes(1);
-
-        const calledUrl = fetch.mock.calls[0][0];
-        expect(calledUrl).toContain("https://test.courses.api.com");
-        expect(calledUrl).toContain("max=20");
-    });
-
-    it("calls the Courses API with the correct base url", async () => {
-        await getServerSideProps(contextWithSearchTerm);
-
-        expect(fetch).toHaveBeenCalledTimes(1);
-
-        const calledUrl = fetch.mock.calls[0][0];
-        expect(calledUrl).toContain(process.env.COURSES_API_BASEURL);
-    });
-
-    it("does not call the Courses API when no search term is entered", async () => {
+    it("does not search for courses when no search term is entered", async () => {
         await getServerSideProps(emptyContext);
 
-        expect(fetch).toHaveBeenCalledTimes(0);
+        expect(searchForCourses).toHaveBeenCalledTimes(0);
     });
 
-    it("calls the Courses API with a search term", async () => {
+    it("calls course search with a search term", async () => {
+        searchForCourses.mockResolvedValue({
+            isSuccessfulSearch: true,
+            searchResponseData: { numberOfMatches: 0, results: [] },
+        });
+
         await getServerSideProps(contextWithSearchTerm);
 
-        expect(fetch).toHaveBeenCalledTimes(1);
-
-        const calledUrl = fetch.mock.calls[0][0];
-        expect(calledUrl).toContain("search=english");
+        expect(searchForCourses).toHaveBeenCalledTimes(1);
+        expect(searchForCourses).toHaveBeenCalledWith("english");
     });
 
-    it("calls the Courses API with a maximum number of results to return", async () => {
-        await getServerSideProps(contextWithSearchTerm);
-
-        expect(fetch).toHaveBeenCalledTimes(1);
-
-        const calledUrl = fetch.mock.calls[0][0];
-        expect(calledUrl).toContain(`max=${process.env.COURSES_API_MAX_RESULTS}`);
-    });
-
-    it("indicates when the Courses API search failed (http error response)", async () => {
-        fetch.mockResponse("{}", { status: 500 });
+    it("indicates when the course search failed", async () => {
+        searchForCourses.mockResolvedValue({
+            isSuccessfulSearch: false,
+            searchResponseData: { numberOfMatches: 0, results: [] },
+        });
 
         const response = await getServerSideProps(contextWithSearchTerm);
 
@@ -118,27 +90,16 @@ describe("getServerSideProps", () => {
         expect(response.props.searchResults).toEqual([]);
     });
 
-    it("indicates when the Courses API search failed (network or other error)", async () => {
-        fetch.mockReject(new Error("can not resolve host"));
+    it("returns the number of matches from the course search", async () => {
+        searchForCourses.mockResolvedValue({
+            isSuccessfulSearch: true,
+            searchResponseData: { numberOfMatches: 1, results: [] },
+        });
 
         const response = await getServerSideProps(contextWithSearchTerm);
 
-        expect(response.props.isSuccessfulSearch).toEqual(false);
-        expect(response.props.numberOfMatches).toEqual(0);
-        expect(response.props.searchResults).toEqual([]);
-    });
+        expect(searchForCourses).toHaveBeenCalledTimes(1);
 
-    it("returns the number of matches from the API", async () => {
-        fetch.mockResponse(
-            JSON.stringify({
-                numberOfMatches: 1,
-                results: [],
-            })
-        );
-
-        const response = await getServerSideProps(contextWithSearchTerm);
-
-        expect(fetch).toHaveBeenCalledTimes(1);
         expect(response.props.isSuccessfulSearch).toEqual(true);
         expect(response.props.numberOfMatches).toEqual(1);
     });
